[Generator] Move GeneratedTypes out and enable nullability.
Moved the code out, also made a small improvement by reducing the number
of lookups are needed to get the GeneratedType.
