+++ src/generator.csproj
@@ -101,6 +101,7 @@
     <Compile Include="..\src\bgen\ExtensionMethods.cs" />
     <Compile Include="..\src\bgen\Filters.cs" />
     <Compile Include="..\src\bgen\GeneratedType.cs" />
+    <Compile Include="..\src\bgen\GeneratedTypes.cs" />
     <Compile Include="..\src\bgen\Generator.cs" />
     <Compile Include="..\src\bgen\MarshalInfo.cs" />
     <Compile Include="..\src\bgen\NamespaceManager.cs" />
